Implement Base.one for StringView
I have a lot of `S where S<:AbstractString` code which uses `one(S)` for the empty string to compare against. I didn't realize until I went to add this that `typemin` also works, but hey, that makes the implementation that much easier, and StringViews gives me a zero-copy drop-in replacement for a string buffer with this addition.